requirements.txt
python-dotenv
google-generativeai
pypdf
